+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.0] - 2024-02-19
+
+### Added
+- Add automatic calculation of numberOfGPULayers option
+- Add formatted output of computed memory details
+
+### Fixed
+- Fix numberOfGPULayers option override
+
 ## [1.9.0] - 2024-02-11
 
 ### Added